third_party/WebKit/LayoutTests/bluetooth/characteristic/writeValue/write-updates-value.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'A regular write request to a writable characteristic ' +
    'should update value.';
let characteristic, fake_characteristic;
let textEncoder = new TextEncoder();
let newValue = textEncoder.encode('foo');

bluetooth_test(() => getMeasurementIntervalCharacteristic()
    .then(_ => ({characteristic, fake_characteristic} = _))
    .then(() => assert_equals(characteristic.value, null))
    .then(() => fake_characteristic.setNextWriteResponse({GATT_SUCCESS}))
    .then(() => characteristic.writeValue(newValue))
    .then(() => assert_array_equals(
      new Uint8Array(characteristic.value.buffer),
      new Uint8Array(newValue.buffer)
    )), test_desc);
</script>
